Add interfaces globally
