+++ addon/helpers/unique-id.ts
@@ -6,18 +6,6 @@ interface Identifiable {
   ref: number | string | Element | object;
 };
 
-function makeId (length = 8) {
-  let result = '';
-  const characters = 'abcdefghijklmnopqrstuvwxyz';
-  const charactersLength = characters.length;
-
-  for (let i = 0; i < length; i++) {
-    result += characters.charAt(Math.floor(Math.random() * charactersLength));
-  }
-
-  return result;
-}
-
 /**
  * Returns a guid for the passed `ref`.
  * If no parameter was passed it will generate a random string and create a guid for it.
@@ -25,7 +13,7 @@ function makeId (length = 8) {
  * @param { any } ref
  */
 export function uniqueId ([], { ref }: { ref: Identifiable }): string {
-  return isPresent(ref) ? guidFor(ref) : guidFor(makeId());
+  return isPresent(ref) ? guidFor(ref) : guidFor({});
 }
 
 export default helper(uniqueId);
