+++ tests/dummy/app/config/environment.d.ts
@@ -1,0 +1,14 @@
+export default config;
+
+/**
+ * Type declarations for
+ *    import config from 'my-app/config/environment'
+ */
+declare const config: {
+  environment: string;
+  modulePrefix: string;
+  podModulePrefix: string;
+  locationType: string;
+  rootURL: string;
+  APP: Record<string, unknown>;
+};
